+++ test/unit/test_config.py
@@ -589,9 +589,9 @@ class TestProject(BaseConfigTest):
         self.assertEqual(project.source_paths, ['models'])
         self.assertEqual(project.macro_paths, ['macros'])
         self.assertEqual(project.data_paths, ['data'])
-        self.assertEqual(project.test_paths, ['test'])
-        self.assertEqual(project.analysis_paths, [])
-        self.assertEqual(project.docs_paths, ['models', 'data', 'snapshots', 'macros'])
+        self.assertEqual(project.test_paths, ['tests'])
+        self.assertEqual(project.analysis_paths, ['analyses'])
+        self.assertEqual(project.docs_paths, ['models', 'data', 'snapshots', 'analyses', 'macros'])
         self.assertEqual(project.asset_paths, [])
         self.assertEqual(project.target_path, 'target')
         self.assertEqual(project.clean_targets, ['target'])
@@ -624,7 +624,7 @@ class TestProject(BaseConfigTest):
             'target-path': 'other-target',
         })
         project = project_from_config_norender(self.default_project_data)
-        self.assertEqual(project.docs_paths, ['other-models', 'data', 'snapshots', 'macros'])
+        self.assertEqual(project.docs_paths, ['other-models', 'data', 'snapshots', 'analyses', 'macros'])
         self.assertEqual(project.clean_targets, ['other-target'])
 
     def test_hashed_name(self):
@@ -636,8 +636,8 @@ class TestProject(BaseConfigTest):
             'source-paths': ['other-models'],
             'macro-paths': ['other-macros'],
             'data-paths': ['other-data'],
-            'test-paths': ['other-test'],
-            'analysis-paths': ['analysis'],
+            'test-paths': ['other-tests'],
+            'analysis-paths': ['other-analyses'],
             'docs-paths': ['docs'],
             'asset-paths': ['other-assets'],
             'target-path': 'other-target',
@@ -706,8 +706,8 @@ class TestProject(BaseConfigTest):
         self.assertEqual(project.source_paths, ['other-models'])
         self.assertEqual(project.macro_paths, ['other-macros'])
         self.assertEqual(project.data_paths, ['other-data'])
-        self.assertEqual(project.test_paths, ['other-test'])
-        self.assertEqual(project.analysis_paths, ['analysis'])
+        self.assertEqual(project.test_paths, ['other-tests'])
+        self.assertEqual(project.analysis_paths, ['other-analyses'])
         self.assertEqual(project.docs_paths, ['docs'])
         self.assertEqual(project.asset_paths, ['other-assets'])
         self.assertEqual(project.target_path, 'other-target')
@@ -1201,9 +1201,9 @@ class TestRuntimeConfigFiles(BaseFileTest):
         self.assertEqual(config.source_paths, ['models'])
         self.assertEqual(config.macro_paths, ['macros'])
         self.assertEqual(config.data_paths, ['data'])
-        self.assertEqual(config.test_paths, ['test'])
-        self.assertEqual(config.analysis_paths, [])
-        self.assertEqual(config.docs_paths, ['models', 'data', 'snapshots', 'macros'])
+        self.assertEqual(config.test_paths, ['tests'])
+        self.assertEqual(config.analysis_paths, ['analyses'])
+        self.assertEqual(config.docs_paths, ['models', 'data', 'snapshots', 'analyses', 'macros'])
         self.assertEqual(config.asset_paths, [])
         self.assertEqual(config.target_path, 'target')
         self.assertEqual(config.clean_targets, ['target'])